refactor: move mutant diff/apply out of __main__ into mutation/diff_apply.py
Remove the need to pass in function to the browser and
move mutation functions out of main.
